ref(profile-chunks): Separate profile chunks pipeline steps, cleanup relay-profiling
